Backend/requirements.txt
fastapi
uvicorn
web3
requests
python-multipart
google-generativeai
